+++ src/mako/chrono/exceptions/ChronoException.php
@@ -1,0 +1,18 @@
+<?php
+
+/**
+ * @copyright Frederic G. Østby
+ * @license   http://www.makoframework.com/license
+ */
+
+namespace mako\chrono\exceptions;
+
+use RuntimeException;
+
+/**
+ * Chrono exception.
+ */
+class ChronoException extends RuntimeException
+{
+
+}